+++ CHANGES.md
@@ -1,6 +1,7 @@
 ## Version 1.3.5
 ## (in development)
 * add picon value to getservices request #849
+* add genre to epg result #847
 
 ## Version 1.3.4
 * Autotimer backup/restore